Options files need to specify symbols, for now.
This is ugly, and was required by the setup that deploys to AWS Lambda, because for some reason the options hashes in memory had keys that were strings, not symbols.  The best fix would be for the options loading code to ensure that they’re symbols, so that the YAML files wouldn’t be so brittle.